+++ imgui-glium-renderer/src/lib.rs
@@ -1,3 +1,6 @@
+pub extern crate glium;
+pub extern crate imgui;
+
 use glium::backend::{Context, Facade};
 use glium::index::{self, PrimitiveType};
 use glium::program::ProgramChooserCreationError;